add user authentication to snippets endpoints

+++ server/routers/snippetRouter.js
@@ -4,16 +4,14 @@ const auth = require("../middleware/auth");
 
 router.get("/", auth, async (req, res) => {
   try {
-    console.log(req.user);
-
-    const snippets = await Snippet.find();
+    const snippets = await Snippet.find({ user: req.user });
     res.json(snippets);
   } catch (error) {
     res.status(500).send();
   }
 });
 
-router.post("/", async (req, res) => {
+router.post("/", auth, async (req, res) => {
   try {
     const { title, description, code } = req.body;
 
@@ -25,10 +23,13 @@ router.post("/", async (req, res) => {
       });
     }
 
+    console.log(req.user);
+
     const newSnippet = new Snippet({
       title,
       description,
       code,
+      user: req.user,
     });
 
     const savedSnippet = await newSnippet.save();
@@ -39,7 +40,7 @@ router.post("/", async (req, res) => {
   }
 });
 
-router.put("/:id", async (req, res) => {
+router.put("/:id", auth, async (req, res) => {
   try {
     const { title, description, code } = req.body;
     const snippetId = req.params.id;
@@ -65,6 +66,10 @@ router.put("/:id", async (req, res) => {
       });
     }
 
+    if (originalSnippet.user.toString() !== req.user) {
+      return res.status(401).json({ errorMessage: "Unauthorized." });
+    }
+
     originalSnippet.title = title;
     originalSnippet.description = description;
     originalSnippet.code = code;
@@ -77,7 +82,7 @@ router.put("/:id", async (req, res) => {
   }
 });
 
-router.delete("/:id", async (req, res) => {
+router.delete("/:id", auth, async (req, res) => {
   try {
     const snippetId = req.params.id;
 
@@ -95,6 +100,10 @@ router.delete("/:id", async (req, res) => {
         errorMessage:
           "No snippet with this ID was found. Please contact the developer.",
       });
+    }
+
+    if (existingSnippet.user.toString() !== req.user) {
+      return res.status(401).json({ errorMessage: "Unauthorized." });
     }
 
     await existingSnippet.delete();
